share mcp mock and life span between ui and api tests
